refactor(callbacks/otel): implement OpenTelemetry tracing for callback lifecycle- Add tracing for OnStartWithStreamInput and OnEndWithStreamOutput- Implement span recording for callback input and output - Refactor state handling to improve trace context management - Move span attribute setting to separate functions for better organization
